BETA: WalkingManager impovements, random respawn interval fix. 	* Support for attaching certain route to certain NPC with certain spawn coordinates in Routes.xml 	* onNodeArrive quest trigger, triggering, when NPC, which is controlled by Walking Manager, arrives next node 	* proper handling for situation, when monster, which is controlled by Walking Manager, is attacked by player; handling for minions of Walking Manager controlled monsters. 	* AI-disabled NPC's, having aggro_range parameter, which is controlled by Walking Manager, can track aggro range enter 	* receiving debug information about walk using //debug command 	* random respawn interval fix - _respawnMinDelay and _respawnMaxDelay are reverted to int.
	* Reported by: UnAfraid (random respawn interval)
	* Reviewed by: UnAfraid, Zoey76

+++ L2J_Server_BETA/java/com/l2jserver/gameserver/instancemanager/RaidBossSpawnManager.java
@@ -188,9 +188,9 @@ public class RaidBossSpawnManager
 		{
 			boss.setRaidStatus(StatusEnum.DEAD);
 			
-			final long respawnMinDelay = (long) (boss.getSpawn().getRespawnMinDelay() * Config.RAID_MIN_RESPAWN_MULTIPLIER);
-			final long respawnMaxDelay = (long) (boss.getSpawn().getRespawnMaxDelay() * Config.RAID_MAX_RESPAWN_MULTIPLIER);
-			final long respawnDelay = Rnd.get(respawnMinDelay, respawnMaxDelay);
+			final int respawnMinDelay = (int) (boss.getSpawn().getRespawnMinDelay() * Config.RAID_MIN_RESPAWN_MULTIPLIER);
+			final int respawnMaxDelay = (int) (boss.getSpawn().getRespawnMaxDelay() * Config.RAID_MAX_RESPAWN_MULTIPLIER);
+			final int respawnDelay = Rnd.get(respawnMinDelay, respawnMaxDelay);
 			final long respawnTime = Calendar.getInstance().getTimeInMillis() + respawnDelay;
 			
 			info.set("currentHP", boss.getMaxHp());
